Moved function definitions in a separate file
